fix everything to long, not just 1/2 of it

+++ src/browsers/playlist3-current-playlist-browser.c
@@ -85,17 +85,17 @@ static void pl3_cp_current_song_changed(GmpcMpdDataModelPlaylist *model2,GtkTree
         }
     }
 }
-static void __real_pl3_total_playtime_changed(GmpcMpdDataModelPlaylist *model, unsigned long long loaded_songs, unsigned long long total_playtime, gpointer user_data)
+static void __real_pl3_total_playtime_changed(GmpcMpdDataModelPlaylist *model, unsigned long loaded_songs, unsigned long total_playtime, gpointer user_data)
 {
      if(mpd_playlist_get_playlist_length(connection)&&loaded_songs)
 
      {
-         unsigned long long total_songs = mpd_playlist_get_playlist_length(connection);
+         unsigned long total_songs = mpd_playlist_get_playlist_length(connection);
          guint playtime = total_playtime*((gdouble)(total_songs/(gdouble)loaded_songs));
          gchar *string = format_time(playtime);
          gchar *mesg = NULL;
 
-         mesg = g_strdup_printf("%llu %s%c %s %s", total_songs, 
+         mesg = g_strdup_printf("%lu %s%c %s %s", total_songs, 
                  ngettext("item", "items", total_songs ), (string[0])?',':' ', string,     
                  (string[0] == 0 || total_songs == loaded_songs)? "":_("(Estimation)"));
 
@@ -112,7 +112,7 @@ static void __real_pl3_total_playtime_changed(GmpcMpdDataModelPlaylist *model, u
      }
 }
 
-static void pl3_total_playtime_changed(GmpcMpdDataModelPlaylist *model, unsigned long long loaded_songs, unsigned long long total_playtime, gpointer user_data)
+static void pl3_total_playtime_changed(GmpcMpdDataModelPlaylist *model, unsigned long loaded_songs, unsigned long total_playtime, gpointer user_data)
 {
     if(pl3_cat_get_selected_browser() == current_playlist_plug.id)
     {
@@ -757,7 +757,7 @@ static void pl3_current_playlist_browser_show_info()
 
 static void pl3_current_playlist_browser_selected(GtkWidget *container)
 {
-    unsigned long long a,b;
+    unsigned long a,b;
     if(pl3_cp_vbox == NULL)
     {
         pl3_current_playlist_browser_init();
